Added export button and moved the component to a different file
